🔨 move sorted series to chart state
